Fix spirv.
