+++ source/slang/slang-ir-resolve-texture-format.h
@@ -1,0 +1,6 @@
+#include "slang-ir.h"
+
+namespace Slang
+{
+    void resolveTextureFormat(IRModule* module);
+}
